pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ghostprompt"
version = "0.1.0"
description = "Versioned prompt registry with semantic search, hashing, and audit trail"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Adam Thomas", email = "adam@ghostlogic.dev" },
]
keywords = ["prompts", "llm", "templates", "versioning", "semantic-search", "ghostlogic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "maelspine>=0.1.1",
    "ghostseal>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/adam-scott-thomas/ghostprompt"
Repository = "https://github.com/adam-scott-thomas/ghostprompt"

[tool.hatch.build.targets.wheel]
packages = ["src/ghostprompt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
